update available steps and register
